+++ server/db/test-data.js
@@ -1,0 +1,77 @@
+const getGameRes = {
+  id: 3480,
+  cover: {
+    id: 3592,
+    image_id: "sgpdlhpeaohxwr6ectsy"
+  },
+  name: "Earthworm Jim",
+  slug: "earthworm-jim",
+  summary:
+    "A crow is chasing a worm named Jim while in outer space Psy-Crow is chasing a renegade ship. The ship's captain has stolen an ultra-high-tech-indestructible-super-space-cyber-suit and Queen Slug-for-a-Butt has ordered Psy-Crow to get it, since it can make her more beautiful than Princess-What's-Her-Name. Psy-Crow blasts the captain and the suit falls to Planet Earth. \n \nBack on earth Jim wonders if he is finally safe when an ultra-high-tech-indestructible-super-space-cyber-suit lands on him. Luckily Jim rests in the neck ring of the suit. Then the space particles begin interacting with Jim, causing a light-speed evolution. Jim soon realizes he is in control of the suit. \n \nJim overhears the Queen's plans for the suit and decides to meet this Princess...",
+  genres: [
+    {
+      id: 5,
+      name: "Shooter"
+    },
+    {
+      id: 8,
+      name: "Platform"
+    },
+    {
+      id: 31,
+      name: "Adventure"
+    }
+  ],
+  platforms: [
+    {
+      id: 5,
+      name: "Wii"
+    },
+    {
+      id: 13,
+      name: "PC DOS"
+    },
+    {
+      id: 19,
+      name: "Super Nintendo Entertainment System (SNES)"
+    },
+    {
+      id: 24,
+      name: "Game Boy Advance"
+    },
+    {
+      id: 29,
+      name: "Sega Mega Drive/Genesis"
+    },
+    {
+      id: 33,
+      name: "Game Boy"
+    },
+    {
+      id: 35,
+      name: "Sega Game Gear"
+    },
+    {
+      id: 36,
+      name: "Xbox Live Arcade"
+    },
+    {
+      id: 39,
+      name: "iOS"
+    },
+    {
+      id: 45,
+      name: "PlayStation Network"
+    },
+    {
+      id: 47,
+      name: "Virtual Console (Nintendo)"
+    },
+    {
+      id: 64,
+      name: "Sega Master System"
+    }
+  ]
+};
+
+module.exports = { getGameRes };
