+++ server/routes/games.js
@@ -47,7 +47,7 @@ router.post("/", (req, res, next) => {
   return igdbApi
     .getGame(igdbId)
     .then(res => {
-      const { name, cover, slug, summary, genres } = res;
+      const { name, cover, slug, summary, genres, platforms } = res;
       const { image_id } = cover;
       const newGame = {
         igdb: {
@@ -57,7 +57,8 @@ router.post("/", (req, res, next) => {
         name,
         coverUrl: `https://images.igdb.com/igdb/image/upload/t_720p/${image_id}.jpg`,
         summary,
-        genres
+        genres,
+        platforms
       };
       return Game.create(newGame);
     })
